- Implement array for passenger & class spinner
